delete and read fix

+++ webapp/src/app/pages/diary/overview/overview.component.ts
@@ -1,7 +1,10 @@
-import { Component, OnInit, ViewChild } from '@angular/core';
+import { Component, OnInit, ViewChild, Input } from '@angular/core';
 import { Record } from '@models/record/';
+import { mocked } from '@models/user';
 import { MatCalendarCellCssClasses, MatCalendar } from '@angular/material/datepicker';
+import { DiaryService } from '@services/diary/diary.service';
 import { AuthService } from '@services/auth/auth.service';
+import { UtilsService } from '@services/utils/utils.service';
 
 function checkDate(utc: string, incomplete: Date): boolean {
 	let date = new Date(utc).toLocaleDateString();
@@ -14,10 +17,10 @@ function checkDate(utc: string, incomplete: Date): boolean {
 	styleUrls: ['./overview.component.sass'],
 })
 export class OverviewComponent implements OnInit {
-	records: Record[] = [];
 	order: string = 'desc';
 	startingDate: Date = new Date('04/15/2020');
 	dateClass: Function = (d: Date): MatCalendarCellCssClasses => {
+		if (!this.records) return;
 		let record = this.records.find((record: Record) => {
 			return checkDate(record.date, d);
 		});
@@ -27,10 +30,14 @@ export class OverviewComponent implements OnInit {
 		return classes;
 	};
 	@ViewChild('calendar') calendar: MatCalendar<Date>;
-	constructor(private auth: AuthService) {}
-
+	records: Record[];
+	constructor(
+		private diaryService: DiaryService,
+		private auth: AuthService,
+		private utils: UtilsService
+	) {}
 	ngOnInit(): void {
-		this.auth.records$.subscribe((records: Record[]) => {
+		this.diaryService.records.subscribe((records: Record[]) => {
 			this.records = records;
 			this.orderNotes();
 			if (this.calendar) this.calendar.updateTodaysDate(); // update calendar view
@@ -38,7 +45,7 @@ export class OverviewComponent implements OnInit {
 	}
 
 	filterWithNotes(records: Record[]): Record[] {
-		return records.filter((record: Record) => (record.notes ? true : false));
+		return records ? records.filter((record: Record) => (record.notes ? true : false)) : [];
 	}
 
 	formatDate(date: string): string {
@@ -46,10 +53,29 @@ export class OverviewComponent implements OnInit {
 	}
 
 	deleteRecord(record: Record) {
-		console.info('delete ', record);
+		console.info('delete', record);
 		let i = this.records.findIndex((r: Record) => r.date == record.date);
-		this.records.splice(i, 1);
-		this.calendar.updateTodaysDate(); // update calendar view
+		if (i !== -1)
+			// if found
+			this.auth
+				.deleteRecord(mocked, record)
+				.then((res: boolean) => {
+					if (res) {
+						this.utils.openSnackBar('Record on date ' + record.date, 'Deleted');
+						this.records.splice(i, 1); // delete from the local copy
+						this.diaryService.records.next(this.records); // send update
+						this.calendar.updateTodaysDate(); // update calendar view
+					} else
+						this.utils.openSnackBar(
+							'Error while inserting new Record',
+							'Please try again 🙏'
+						);
+				})
+				.catch(err => {
+					console.error(err);
+					this.utils.openSnackBar('Something went wrong', '💀💀💀');
+				});
+		else console.error('Invalid index');
 	}
 
 	orderNotes() {
